--- a/.actrc
+++ b/.actrc
@@ -1,0 +1,21 @@
+# Act configuration for testing GitHub Actions locally
+# https://github.com/nektos/act
+
+# Default runner images
+-P ubuntu-latest=catthehacker/ubuntu:act-latest
+-P ubuntu-22.04=catthehacker/ubuntu:act-22.04
+-P ubuntu-20.04=catthehacker/ubuntu:act-20.04
+
+# macOS runners (use Ubuntu as act doesn't support real macOS)
+-P macos-latest=catthehacker/ubuntu:act-latest
+-P macos-13=catthehacker/ubuntu:act-latest
+-P macos-12=catthehacker/ubuntu:act-latest
+
+# Container architecture
+--container-architecture linux/amd64
+
+# Reuse containers for faster runs
+--reuse
+
+# Use Docker by default
+--container-daemon-socket -