--- a/Brewfile
+++ b/Brewfile
@@ -1,0 +1,51 @@
+# ===========================
+# TAPS FOR SPECIALIZED TOOLS
+# ===========================
+tap "kyoh86/tap"  # For richgo (enhanced go test output)
+
+# ===========================
+# CORE FILE SEARCH & NAVIGATION TOOLS
+# ===========================
+brew "fd"              # Fast find alternative (used in fzf commands)
+brew "fzf"             # Fuzzy finder (heavily used in aliases)
+brew "ripgrep"         # Fast grep alternative (has config file)
+brew "bat"             # Cat with syntax highlighting (used in fzf previews)
+
+# ===========================
+# DIRECTORY & FILE MANAGEMENT
+# ===========================
+brew "lsd"             # Modern ls replacement (used in aliases)
+brew "eza"             # Modern ls replacement (used in fzf previews)
+brew "zoxide"          # Smart cd replacement (used in zz alias)
+brew "tree"            # Directory tree viewer (used as fallback)
+brew "broot"           # Interactive directory navigation (has config)
+brew "stow"            # Dotfiles symlink manager (used in Makefile)
+
+# ===========================
+# DEVELOPMENT TOOLS
+# ===========================
+brew "neovim"          # Editor (aliased as vim, vi, v)
+brew "docker"          # Containerization (has aliases and functions)
+brew "kubernetes-cli"  # Kubectl (lazy loaded in zshrc)
+
+# ===========================
+# SYSTEM MONITORING
+# ===========================
+brew "btop"            # System monitor (has config directory)
+
+# ===========================
+# ZSH ENHANCEMENTS
+# ===========================
+brew "powerlevel10k"   # ZSH theme
+
+# ===========================
+# LANGUAGE-SPECIFIC TOOLS
+# ===========================
+brew "kyoh86/tap/richgo"  # Enhanced go test output (aliased as 'go')
+
+# ===========================
+# APPLICATIONS & TOOLS
+# ===========================
+cask "ghostty"                # Terminal emulator (extensive config in dotfiles)
+cask "git-credential-manager" # Git credential helper (referenced in .gitconfig)
+cask "miniconda"              # Python environment manager (lazy loaded in zshrc)