--- a/broot/.config/broot/conf.hjson
+++ b/broot/.config/broot/conf.hjson
@@ -24,7 +24,7 @@
 # you can launch broot with `br -H`).
 # A popular flag is the `g` one which displays git related info.
 #
-# default_flags:
+#default_flags: -h
 
 ###############################################################
 # Terminal's title
@@ -100,7 +100,7 @@ show_selection_mark: true
 # Special paths
 # If some paths must be handled specially, uncomment (and change
 # this section as per the examples)
-# Setting "list":"never" on a dir prevents broot from looking at its 
+# Setting "list":"never" on a dir prevents broot from looking at its
 #  children when searching, unless the dir is the selected root.
 # Setting "sum":"never" on a dir prevents broot from looking at its
 #  children when computing the total size and count of files.
@@ -130,7 +130,7 @@ special_paths: {
 # If you want the escape key to quit broot when there's nothing
 # to cancel (for example when you just opened broot), uncomment
 # this parameter
-# 
+#
 # quit_on_last_cancel: true
 
 ###############################################################
@@ -201,11 +201,11 @@ content_search_max_file_size: 10MB
 #
 # If you want to use advanced keyboard shortcuts in Kitty
 # compatible terminals (Kitty, Wezterm), set this to true.
-# 
-# This makes it possible to use shortcuts like 'space-n', 
+#
+# This makes it possible to use shortcuts like 'space-n',
 # 'ctrl-alt-a-b', 'shift-space', etc.
 #
-# enable_kitty_keyboard: false
+enable_kitty_keyboard: true
 
 ###############################################################
 # lines around matching line in filtered preview
@@ -252,7 +252,7 @@ preview_transformers: [
     // # In this example, the command refers to neither the input nor the output,
     // # so broot pipes them to the stdin and stdout of the jq process
     // {
-    //     input_extensions: [ "json" ] 
+    //     input_extensions: [ "json" ]
     //     output_extension: json
     //     mode: text
     //     command: [ "jq" ]
@@ -265,7 +265,7 @@ preview_transformers: [
 # While it's possible to have all configuration in one file,
 # it's more convenient to split it in several ones.
 # Importing also allows to set a condition on the terminal's
-# color, which makes it possible to have a different skin 
+# color, which makes it possible to have a different skin
 # chosen when your terminal has a light background and when
 # it has a light one.
 imports: [