--- a/nvim/.config/nvim/lua/configs/indent-blankline.lua
+++ b/nvim/.config/nvim/lua/configs/indent-blankline.lua
@@ -1,0 +1,4 @@
+require("indent_blankline").setup {
+  indent_blankline_char = "│",
+  show_current_context = true,
+}