… .claude-nvim/settings.local.json
… .claude-zsh/settings.local.json
… .github/workflows/README.md
… .github/workflows/claude-code-review.yml
… .github/workflows/claude.yml
… .github/workflows/comprehensive-test.yml
… .github/workflows/fast-test.yml
… .github/workflows/test-simple.yml
… .github/workflows/test.yml
… README.md
… atuin/.config/atuin/config.toml
… broot/README.md
… broot/broot_tutorial.md
… btop/README.md
… ghostty/.config/ghostty/config
… ghostty/.config/ghostty/profiles.conf
… ghostty/.config/ghostty/themes.conf
… ghostty/README.md
… git/.gitconfig
… git/.gitconfig.d/aliases
… git/.gitconfig.d/tools
… git/README.md
… go/README.md
… install.sh
… kitty/.config/kitty/current-theme.conf
… kitty/.config/kitty/kitty.conf
… kitty/.config/kitty/macos-shortcuts.conf
… kitty/.config/kitty/reading-optimizations.conf
… kitty/.config/kitty/theme-specific-settings.conf
… kitty/.config/kitty/themes/github-light.conf